renaming files and cleanup
